♻️ [Refactor] requestJoin 함수 분리
