+++ frontend/src/pages/Join/Join.jsx
@@ -1,12 +1,23 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
 import axios from 'axios';
-import { Button, Container, Input, Form } from '../../components';
+import { Button, Container, Input, Form, Title } from '../../components';
 import { useNavigate } from 'react-router-dom';
+import { requestJoin } from '@/api/requestJoin';
 
-const Ul = styled.ul`
-  background-color: lightgreen;
-`
+
+const ButtonWrapper = styled.div`
+  width: 100%;
+  padding: 50px 0 0;
+  text-align: center;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 20px;
+  color: #6B4EFF;
+`;
+
+
 const Li = styled.li`
   display: flex;
   height: 60px;
@@ -37,52 +48,34 @@ function Join() {
   const handlePasswordChange = (e) => setPassword(e.target.value);
   const handleRePasswordChange = (e) => setRePassword(e.target.value);
   const handleRoleChange = (e) => setRole(e.target.value);
-  
-  const navigate = useNavigate();
 
+  const Join = requestJoin();
+
+  const joinData = {
+    email,
+    name,
+    loginPassword: password,
+    role
+  };
 
   // 회원가입 폼 제출 요청 
   const handleSubmitForm = async (event) => {
     event.preventDefault();
-
 
     if(password !== rePassword) {
       alert('비밀번호가 일치하지 않습니다.');
       return;
     }
 
-    const joinData = {
-      email,
-      name,
-      loginPassword: password,
-      role
-    };
-    
-    console.log(joinData);
-
-    try {
-      const response = await axios.post('/api/join', joinData);
-      alert('회원가입 성공: ' + response.data.message);
-      navigate('/login');
-    } catch (error) {
-      if (error.response) {
-        // 요청이 이루어졌으나 서버가 2xx 범위가 아닌 상태 코드로 응답
-        alert('회원가입 실패: ' + error.response.data.message);
-      } else if (error.request) {
-        // 요청이 이루어 졌으나 응답을 받지 못함
-        alert('회원가입 요청 실패: 서버에서 응답이 없습니다.');
-      } else {
-        // 요청을 설정하는 중에 문제가 발생함
-        alert('회원가입 요청 중 오류 발생: ' + error.message);
-      }
-    }
+    Join(joinData);
   }
 
 
   return (
     <Container as="section" align="center" width="widthMedium">
+      <Title as="h2" titleStyle="XXL" align="center">Join DND</Title>
       <Form onSubmit={handleSubmitForm}>
-        <Ul>
+        <ul>
           <Li>
             <Label htmlFor="name">이름</Label>
             <Input 
@@ -152,7 +145,7 @@ function Join() {
               관리자
             </Label>
           </Li>
-        </Ul>
+        </ul>
         <Button type="submit">회원가입하기</Button>
       </Form>
     </Container>
